TWRSS/requirements.txt
anyio==4.2.0
certifi==2024.2.2
exceptiongroup==1.2.0
fake-useragent==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
twikit==1.1.26
typing_extensions==4.9.0
zipp==3.17.0
